requirements.txt
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
ftfy==6.3.1
regex==2024.11.6
tqdm==4.67.1
triton==3.2.0
tensorboard==2.20.0
datetime==5.5
pillow==11.0.0
numpy==1.24.4